Add reusable license key clipboard component.
