+++ client/my-sites/checkout/checkout-thank-you/licensing-thank-you-manual-activation-license-key.tsx
@@ -1,28 +1,23 @@
 import { useTranslate } from 'i18n-calypso';
-import { FC, useCallback, useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { FC } from 'react';
+import { useSelector } from 'react-redux';
 import licensingActivationPluginBanner from 'calypso/assets/images/jetpack/licensing-activation-plugin-banner.svg';
 import QueryProductsList from 'calypso/components/data/query-products-list';
 import LicensingActivation from 'calypso/components/jetpack/licensing-activation';
-import useUserLicenseByReceiptQuery from 'calypso/data/jetpack-licensing/use-user-license-by-receipt-query';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
-import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import {
 	isProductsListFetching as getIsProductListFetching,
 	getProductName,
 } from 'calypso/state/products-list/selectors';
-import JetpackLicenseKeyClipboard from './jetpack-license-key-clipboard';
+import JetpackLicenseKeyClipboard, {
+	JetpackLicenseKeyProps,
+} from './jetpack-license-key-clipboard';
 
-interface Props {
-	productSlug: string | 'no_product';
-	receiptId: number;
-}
-
-const LicensingActivationInstructions: FC< Props > = ( { productSlug, receiptId } ) => {
+const LicensingActivationInstructions: FC< JetpackLicenseKeyProps > = ( {
+	productSlug,
+	receiptId,
+} ) => {
 	const translate = useTranslate();
-	const dispatch = useDispatch();
-
-	const [ isCopied, setCopied ] = useState( false );
 
 	const hasProductInfo = productSlug !== 'no_product';
 
@@ -31,34 +26,6 @@ const LicensingActivationInstructions: FC< Props > = ( { productSlug, receiptId 
 	);
 
 	const isProductListFetching = useSelector( ( state ) => getIsProductListFetching( state ) );
-
-	const {
-		data: dataLicense,
-		isError: isErrorFetchingLicense,
-		isLoading: isLoadingLicense,
-	} = useUserLicenseByReceiptQuery( receiptId );
-
-	const licenseKey =
-		! isErrorFetchingLicense && ! isLoadingLicense && dataLicense
-			? dataLicense[ 0 ].licenseKey
-			: '';
-
-	const onCopy = useCallback( () => {
-		setCopied( true );
-		dispatch(
-			recordTracksEvent( 'calypso_siteless_checkout_manual_activation_license_key_copy', {
-				product_slug: productSlug,
-				license_key: licenseKey,
-			} )
-		);
-	}, [ dispatch, licenseKey, productSlug ] );
-
-	useEffect( () => {
-		if ( isCopied ) {
-			const confirmationTimeout = setTimeout( () => setCopied( false ), 4000 );
-			return () => clearTimeout( confirmationTimeout );
-		}
-	}, [ isCopied ] );
 
 	return (
 		<>
@@ -91,13 +58,7 @@ const LicensingActivationInstructions: FC< Props > = ( { productSlug, receiptId 
 					) }
 				</p>
 
-				<JetpackLicenseKeyClipboard
-					copied={ isCopied }
-					disabled={ isErrorFetchingLicense || isLoadingLicense }
-					licenseKey={ licenseKey }
-					loading={ isLoadingLicense }
-					onCopy={ onCopy }
-				/>
+				<JetpackLicenseKeyClipboard productSlug={ productSlug } receiptId={ receiptId } />
 			</LicensingActivation>
 		</>
 	);
